+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## [0.1.2] - Added support for invisible moneybutton
+## [0.1.1] - swipe future now returns nested Map instead of js object
+
+## [0.1.0] - Added support for invisible moneybutton
 
 ## [0.0.1] - An experimental moneybutton plugin for flutter web